Make sure that selected run mask pixel, veg map pixel value and CMT type are all copasetic. The function calls stop() if there is anything inconsistent, for example more tha one pixel is enabled in the run mask, or the enabled pixel's vegetation type does not match the vegetation/community type of the chosen PFTs.

Make sure that selected run mask pixel, veg map pixel value and CMT type are all copasetic. The function calls stop() if there is anything inconsistent, for example more tha one pixel is enabled in the run mask, or the enabled pixel's vegetation type does not match the vegetation/community type of the chosen PFTs.

Usage

enforce.runmask.cmt.vegmap.harmony(siteDataPath, rundir, cmtnum)

Arguments

siteDataPath

is the path to the folder where we expect to find the dvmdostem input data files.

rundir

is the path to the local running directory where customized files (config, parameters, runmask etc) are copied to.

cmtnum

is the community type (vegetation type) that should be used for the run. Based on the chosen PFT, and required to look up the correct parameters in the parameter files.

Value

none

Author

Tobey Carman