Write data into PEcAn shaped output file.

Write data into PEcAn shaped output file.

Usage

write.data2pecan.file(
  y_starts,
  outdir,
  pecan_requested_vars,
  monthly_dvmdostem_outputs,
  yearly_dvmdostem_outputs,
  px_Y,
  px_X
)

Arguments

y_starts

a list of years, i.e.: 1901, 1902, 1903, etc.

outdir

a path to the location where were we will look for dvmdostem outputs and write PEcAn outputs.

pecan_requested_vars

comma separated string listing the variables to process (PEcAn names).

monthly_dvmdostem_outputs

list of files available from dvmdostem at monthly resolution.

yearly_dvmdostem_outputs

list of files available from dvmdostem at yearly resolution.

px_Y

the pixel offset, Y (latitude) dimension.

px_X

the pixel offset, X (longitude) dimension.

Author

Tobey Carman