Setup outputs to be generated by dvmdostem and analyzed by PEcAn.

Setup the output variables that dvmdostem will generate and PEcAn will analyze. This function handles the interplay between output variables and output spec file. There are custom tags in the <model> section of the pecan xml file for dvmdostem that allow the user to specifiy a list of variables to generate and process, and a custom path to an output spec file. The need for a custom path to an output spec file is mostly gone with the addition of the variable list, but will be left in place in case there is a future need for it. This function looks at the client supplied settings for output spec path and variable list and returns the path to the final run specific output spec file and the list of variables to process. The run specific output spec file is copied into place and adjusted using the dvmdostem script for working with output spec files.

Usage

setup.outputs.dvmdostem(
  dvmdostem_calibration,
  pecan_requested_outputs,
  dvmdostem_output_spec,
  run_directory,
  run_id,
  appbinary_path
)

Arguments

dvmdostem_calibration

a string with 'yes' or 'YES'

pecan_requested_outputs

a space separated string of variables to process or NULL.

dvmdostem_output_spec

a path to a custom output spec file or NULL.

run_directory

a path to the direcotory containing the PEcAn run.

run_id

the identifier for this individual run (usually a 10 digit number).

appbinary_path

path to the dvmdostem application.

Value

Vector containing path to the run specific output spec file and the final space separated string of out variables to pecanify.

Author

Tobey Carman